requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
future==0.18.2
idna==2.9
oauthlib==3.1.0
pkg-resources==0.0.0
psycopg2==2.8.4
python-dotenv==0.11.0
python-twitter==3.5
requests==2.23.0
requests-oauthlib==1.3.0
soupsieve==2.0
urllib3==1.25.8
